Support for delta-format in input file
Modified the loadEpisodes method in main so that it can load a file
where the episodes are in a delta format wherein each episode (besides
the first) is specified in terms of which WMEs add/remove from the
previous episode.  See log_eaters_changesonly2.txt for an example of
this format.