added functionality to calculate success
calculate success now has a third data point that finds the most 
similar hashcode in the hashcodelist then calculates whether or not
the most similar hashcode belongs to the correct episode.  To do this,
we added a HashCodeEpList which keeps track of the location in the 
episodeList of each hashcode in the hashcodelist.