fixed JUnit test for future
testGenerateHashFn is not currently passing but that is because it is
set up to pass when the resistance to change positions is implemented
for the SweetSpotHashFn, which is not implemented as of now.  But once
we add that functionality, the test should pass.

Side note:  the GAHashFn test is passing SOMETIMES and not passing other
times, we suspect that there may be a bug somewhere in that code.